+++ apps/storybook/.storybook/main.ts
@@ -10,6 +10,7 @@ const root = path.resolve(dirname, "../../..");
 const config: StorybookConfig = {
   stories: ["../stories/**/*.stories.@(ts|tsx)"],
   addons: ["@storybook/addon-essentials"],
+  staticDirs: ["../public"],
   framework: {
     name: "@storybook/react-vite",
     options: {},